fix: added explicit casting in prepend_id to avoid weird overflow

+++ src/csp_id.c
@@ -125,14 +125,16 @@ static void csp_id1_setup_rx(csp_packet_t * packet) {
 #define CSP_ID2_HEADER_SIZE 6
 
 static void csp_id2_prepend(csp_packet_t * packet) {
+	//NOTE: I had to add the explicit casting to the last three fields or something
+	// 		weird was happening on the MSP430
 
 	/* Pack into 64-bit using host endian */
 	uint64_t id2 = ((((uint64_t)packet->id.pri) << CSP_ID2_PRIO_OFFSET) |
 					(((uint64_t)packet->id.dst) << CSP_ID2_DST_OFFSET) |
 					(((uint64_t)packet->id.src) << CSP_ID2_SRC_OFFSET) |
-					(packet->id.dport << CSP_ID2_DPORT_OFFSET) |
-					(packet->id.sport << CSP_ID2_SPORT_OFFSET) |
-					(packet->id.flags << CSP_ID2_FLAGS_OFFSET));
+					(((uint64_t)packet->id.dport) << CSP_ID2_DPORT_OFFSET) |
+					(((uint64_t)packet->id.sport) << CSP_ID2_SPORT_OFFSET) |
+					(((uint64_t)packet->id.flags) << CSP_ID2_FLAGS_OFFSET));
 
 	/* Convert to big / network endian:
 	 * We first shift up the 48 bit header to most significant end of the 64-bit */
